bleh !!

+++ Common/GlobalItems/AequusItem.Tooltips.cs
@@ -235,7 +235,7 @@ namespace Aequus.Items
         public TooltipLine GetPriceTooltipLine(Player player, Item item)
         {
             player.GetItemExpectedPrice(item, out var calcForSelling, out var calcForBuying);
-            long value = item.isAShopItem || item.buyOnce ? calcForBuying : calcForSelling;
+            int value = item.isAShopItem || item.buyOnce ? calcForBuying : calcForSelling;
             if (item.shopSpecialCurrency != -1)
             {
                 string[] text = new string[1];
